+++ src/direction/hex_direction.rs
@@ -23,15 +23,15 @@ use crate::{DiagonalDirection, HexOrientation};
 ///
 /// ## Operations
 ///
-/// Directions can be rotated:
-///  - *clockwise* with:
+/// Directions can be:
+///  - rotated *clockwise* with:
 ///     - [`Self::right`] and [`Self::rotate_right`]
 ///     - The shift right `>>` operator
-///  - *counter clockwise* with:
+///  - rotated *counter clockwise* with:
 ///     - [`Self::left`] and [`Self::rotate_left`]
 ///     - The shift left `<<` operator
-///
-/// And negated using the minus `-` operator.
+///  - negated using the minus `-` operator
+///  - multiplied by an `i32`, returning a [`Hex`](crate::Hex) vector
 ///
 /// Example:
 /// ```rust